Move PlacementOptions back
