+++ services/placement/options.go
@@ -20,8 +20,16 @@
 
 package placement
 
+import (
+	"github.com/m3db/m3cluster/services"
+	"github.com/m3db/m3x/instrument"
+)
+
 const (
 	defaultMaxStepSize = 3
+	defaultIsSharded   = true
+	// By default partial replace should be allowed for better distribution.
+	defaultAllowPartialReplace = true
 )
 
 type deploymentOptions struct {
@@ -39,5 +47,97 @@ func (o deploymentOptions) MaxStepSize() int {
 
 func (o deploymentOptions) SetMaxStepSize(stepSize int) DeploymentOptions {
 	o.maxStepSize = stepSize
+	return o
+}
+
+type placementOptions struct {
+	looseRackCheck      bool
+	allowPartialReplace bool
+	isSharded           bool
+	isMirrored          bool
+	isStaged            bool
+	iopts               instrument.Options
+	validZone           string
+	dryrun              bool
+}
+
+// NewOptions returns a default services.PlacementOptions.
+func NewOptions() services.PlacementOptions {
+	return placementOptions{
+		allowPartialReplace: defaultAllowPartialReplace,
+		isSharded:           defaultIsSharded,
+		iopts:               instrument.NewOptions(),
+	}
+}
+
+func (o placementOptions) LooseRackCheck() bool {
+	return o.looseRackCheck
+}
+
+func (o placementOptions) SetLooseRackCheck(looseRackCheck bool) services.PlacementOptions {
+	o.looseRackCheck = looseRackCheck
+	return o
+}
+
+func (o placementOptions) AllowPartialReplace() bool {
+	return o.allowPartialReplace
+}
+
+func (o placementOptions) SetAllowPartialReplace(allowPartialReplace bool) services.PlacementOptions {
+	o.allowPartialReplace = allowPartialReplace
+	return o
+}
+
+func (o placementOptions) IsSharded() bool {
+	return o.isSharded
+}
+
+func (o placementOptions) SetIsSharded(sharded bool) services.PlacementOptions {
+	o.isSharded = sharded
+	return o
+}
+
+func (o placementOptions) IsMirrored() bool {
+	return o.isMirrored
+}
+
+func (o placementOptions) SetIsMirrored(v bool) services.PlacementOptions {
+	o.isMirrored = v
+	return o
+}
+
+func (o placementOptions) IsStaged() bool {
+	return o.isStaged
+}
+
+func (o placementOptions) SetIsStaged(v bool) services.PlacementOptions {
+	o.isStaged = v
+	return o
+}
+
+func (o placementOptions) Dryrun() bool {
+	return o.dryrun
+}
+
+func (o placementOptions) SetDryrun(d bool) services.PlacementOptions {
+	o.dryrun = d
+	return o
+}
+
+func (o placementOptions) InstrumentOptions() instrument.Options {
+	return o.iopts
+}
+
+func (o placementOptions) SetInstrumentOptions(iopts instrument.Options) services.PlacementOptions {
+	o.iopts = iopts
+	return o
+}
+
+func (o placementOptions) ValidZone() string {
+	return o.validZone
+}
+
+func (o placementOptions) SetValidZone(z string) services.PlacementOptions {
+	o.validZone = z
 	return o
 }
