+++ services/services.go
@@ -30,15 +30,11 @@ import (
 	metadataproto "github.com/m3db/m3cluster/generated/proto/metadata"
 	placementproto "github.com/m3db/m3cluster/generated/proto/placement"
 	"github.com/m3db/m3cluster/shard"
-	"github.com/m3db/m3x/instrument"
 )
 
 const (
 	defaultLeaderTimeout = 10 * time.Second
 	defaultResignTimeout = 10 * time.Second
-	defaultIsSharded     = true
-	// By default partial replace should be allowed for better distribution.
-	defaultAllowPartialReplace = true
 )
 
 var (
@@ -410,96 +406,4 @@ func (opts namespaceOpts) MetadataNamespace() string {
 func (opts namespaceOpts) SetMetadataNamespace(v string) NamespaceOptions {
 	opts.metadata = v
 	return opts
-}
-
-type placementOptions struct {
-	looseRackCheck      bool
-	allowPartialReplace bool
-	isSharded           bool
-	isMirrored          bool
-	isStaged            bool
-	iopts               instrument.Options
-	validZone           string
-	dryrun              bool
-}
-
-// NewPlacementOptions returns a default PlacementOptions
-func NewPlacementOptions() PlacementOptions {
-	return placementOptions{
-		allowPartialReplace: defaultAllowPartialReplace,
-		isSharded:           defaultIsSharded,
-		iopts:               instrument.NewOptions(),
-	}
-}
-
-func (o placementOptions) LooseRackCheck() bool {
-	return o.looseRackCheck
-}
-
-func (o placementOptions) SetLooseRackCheck(looseRackCheck bool) PlacementOptions {
-	o.looseRackCheck = looseRackCheck
-	return o
-}
-
-func (o placementOptions) AllowPartialReplace() bool {
-	return o.allowPartialReplace
-}
-
-func (o placementOptions) SetAllowPartialReplace(allowPartialReplace bool) PlacementOptions {
-	o.allowPartialReplace = allowPartialReplace
-	return o
-}
-
-func (o placementOptions) IsSharded() bool {
-	return o.isSharded
-}
-
-func (o placementOptions) SetIsSharded(sharded bool) PlacementOptions {
-	o.isSharded = sharded
-	return o
-}
-
-func (o placementOptions) IsMirrored() bool {
-	return o.isMirrored
-}
-
-func (o placementOptions) SetIsMirrored(v bool) PlacementOptions {
-	o.isMirrored = v
-	return o
-}
-
-func (o placementOptions) IsStaged() bool {
-	return o.isStaged
-}
-
-func (o placementOptions) SetIsStaged(v bool) PlacementOptions {
-	o.isStaged = v
-	return o
-}
-
-func (o placementOptions) Dryrun() bool {
-	return o.dryrun
-}
-
-func (o placementOptions) SetDryrun(d bool) PlacementOptions {
-	o.dryrun = d
-	return o
-}
-
-func (o placementOptions) InstrumentOptions() instrument.Options {
-	return o.iopts
-}
-
-func (o placementOptions) SetInstrumentOptions(iopts instrument.Options) PlacementOptions {
-	o.iopts = iopts
-	return o
-}
-
-func (o placementOptions) ValidZone() string {
-	return o.validZone
-}
-
-func (o placementOptions) SetValidZone(z string) PlacementOptions {
-	o.validZone = z
-	return o
 }
