+++ services/services_test.go
@@ -24,7 +24,6 @@ import (
 	"testing"
 
 	placementproto "github.com/m3db/m3cluster/generated/proto/placement"
-	"github.com/m3db/m3x/instrument"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -46,39 +45,6 @@ func TestNamespaceOptions(t *testing.T) {
 	opts = opts.SetPlacementNamespace("p").SetMetadataNamespace("m")
 	assert.Equal(t, "p", opts.PlacementNamespace())
 	assert.Equal(t, "m", opts.MetadataNamespace())
-}
-
-func TestPlacementOptions(t *testing.T) {
-	o := NewPlacementOptions()
-	assert.False(t, o.LooseRackCheck())
-	assert.True(t, o.AllowPartialReplace())
-	assert.True(t, o.IsSharded())
-	assert.False(t, o.Dryrun())
-	assert.False(t, o.IsMirrored())
-	assert.False(t, o.IsStaged())
-	assert.Equal(t, instrument.NewOptions(), o.InstrumentOptions())
-
-	o = o.SetLooseRackCheck(true)
-	assert.True(t, o.LooseRackCheck())
-
-	o = o.SetAllowPartialReplace(false)
-	assert.False(t, o.AllowPartialReplace())
-
-	o = o.SetIsSharded(false)
-	assert.False(t, o.IsSharded())
-
-	o = o.SetDryrun(true)
-	assert.True(t, o.Dryrun())
-
-	o = o.SetIsMirrored(true)
-	assert.True(t, o.IsMirrored())
-
-	o = o.SetIsStaged(true)
-	assert.True(t, o.IsStaged())
-
-	iopts := instrument.NewOptions().SetMetricsSamplingRate(0.5)
-	o = o.SetInstrumentOptions(iopts)
-	assert.Equal(t, iopts, o.InstrumentOptions())
 }
 
 func TestConvertBetweenProtoAndService(t *testing.T) {
